migrate more panel code into TypeScript
