+++ src/panels/alarm-table/sass/_severity-themes.scss
@@ -1,7 +1,7 @@
 @import '~opennms-style/opennms';
 
 // If you add a theme here, you must also give it an entry in
-// alarm-table/editor.js for the editor drop-down.
+// alarm-table/editor.ts for the editor drop-down.
 
 $severity-themes: (
   'helm': (
